+++ CHANGELOG.md
@@ -9,10 +9,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 ### Added
 
 - Added azureAI sample files
+- Introduced attachment table that allows to work with static files. Export/import, path config and other support included
 
 ### Changed
 
 ### Fixed
+- Fixed incorrect content being generated for external blocks in Designer output
 
 ## [17.0.14] - 2026-02-03